Added explosive barrel. Changes to Explosion script. Moved files around into proper directories.
