+++ Assets/Materials/LevelObjects/barrel_explode.mat.meta
@@ -1,7 +1,8 @@
 fileFormatVersion: 2
-guid: 80504c3389227964891cff6e9930b21a
-PrefabImporter:
+guid: 7a37012dbc72dbb46b96a58c85c455b1
+NativeFormatImporter:
   externalObjects: {}
+  mainObjectFileID: 2100000
   userData: 
   assetBundleName: 
   assetBundleVariant:
