+++ Assets/Models/Objects.meta
@@ -1,6 +1,7 @@
 fileFormatVersion: 2
-guid: a698c97272f631c4ea22475ca9305454
-PrefabImporter:
+guid: f3edcb707e81c284c9fce0f1a7707962
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
