+++ Assets/Prefabs/Effects/explosion_prefab.prefab
@@ -668,8 +668,8 @@ ParticleSystem:
     radiusThickness: 1
     donutRadius: 0.2
     m_Position: {x: 0, y: 0, z: 0}
-    m_Rotation: {x: -90, y: 0, z: 0}
-    m_Scale: {x: 1, y: 1, z: 1}
+    m_Rotation: {x: 270, y: -0, z: 0}
+    m_Scale: {x: 0.85070527, y: 0.50241375, z: 1}
     placementMode: 0
     m_MeshMaterialIndex: 0
     m_MeshNormalOffset: 0
@@ -4917,8 +4917,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 947baa0e7bfabd14c8680b18887d3a70, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  lifeTime: 0.2
-  radius: 5
+  damage: 4
   fragments: {fileID: 7095177697174510587}
   blast: {fileID: 7785448750215647882}
 --- !u!1 &8024243766073461282
