+++ Assets/Prefabs/LevelObjects/3x1.prefab
@@ -122,6 +122,7 @@ MonoBehaviour:
   - {fileID: 0}
   debrisCount: 5
   health: 20
+  deathEffect: {fileID: 0}
 --- !u!1001 &3008513061943224443
 PrefabInstance:
   m_ObjectHideFlags: 0
@@ -210,6 +211,7 @@ MonoBehaviour:
   - {fileID: 0}
   debrisCount: 5
   health: 20
+  deathEffect: {fileID: 0}
 --- !u!1001 &7232775058849473144
 PrefabInstance:
   m_ObjectHideFlags: 0
@@ -293,6 +295,7 @@ MonoBehaviour:
   - {fileID: 0}
   debrisCount: 5
   health: 20
+  deathEffect: {fileID: 0}
 --- !u!4 &8437087151980003305 stripped
 Transform:
   m_CorrespondingSourceObject: {fileID: 1245654901351030161, guid: d02d665ce51ffbb479667f5532a64a64, type: 3}
