+++ Assets/Prefabs/LevelObjects/Obstacles.meta
@@ -1,6 +1,7 @@
 fileFormatVersion: 2
-guid: c346f0f36ca64384f894082922960a6c
-PrefabImporter:
+guid: 3d2a79ceac3fd7341b42fee73d09b97b
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
