+++ Assets/Prefabs/LevelObjects/Obstacles/barrel_explode.prefab.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 5afe338122e2add4d986103808cab9f3
+guid: b62de1747513a1249aaf0bbd6ba35758
 PrefabImporter:
   externalObjects: {}
   userData:
