+++ Assets/Textures/Objects.meta
@@ -1,6 +1,7 @@
 fileFormatVersion: 2
-guid: 6de9bb50f45e94d4cba78a46c8ed6b01
-PrefabImporter:
+guid: 5a9e628359f29904ebc820c75e465dfe
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
