+++ Assets/Textures/Objects/Barrel.meta
@@ -1,6 +1,7 @@
 fileFormatVersion: 2
-guid: 51cf19cffc8106f4f8c3d532f8c25623
-PrefabImporter:
+guid: 4c81b454cbaff74429a0a9086562f150
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
